requirements.txt
loguru
spotipy
python-dotenv
requests
pillow
